# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CodingPlatform Documentation",
        "description": "CodingPlatform Documentation",
        "contact": {
            "name": "API document",
            "email": "s1101778@gm.pu.edu.tw"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "CodingPlatform API Server"
        }
    ],
    "paths": {
...
        "/forum/post": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Post a video",
                "description": "Post a video with the given details",
                "operationId": "postVideo",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "serial": {
                                        "description": "serial",
                                        "type": "string"
                                    },
                                    "video_url": {
                                        "description": "Youtube video url",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "Video content",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful post video",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Post ID",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    }
}